Charlie OS Python Edition V1.0
Charlie OS Python Edition  V1.0 Is The First Version Of  Charlie OS Python Edition That Has A Linux-Style Terminal As The Desktop And Only Has A Help And Shutdown. Needs Python 3.7 To Work. To Get Help You Need To Type Help All. If You Want To Help Develop Then Do Not Commit To The Master Branch, Commit To The Helping Hand Branch.